skills/nanobanana-milan-demo/scripts/requirements.txt
google-genai>=0.2.0